+++ src/runtime/src/core/iterator/generators.rs
@@ -1,6 +1,6 @@
 use crate::{
     value_iterator::{KotoIterator, ValueIterator, ValueIteratorOutput as Output},
-    Value,
+    CallArgs, Value, Vm,
 };
 
 /// An iterator that repeatedly yields the same value
@@ -71,6 +71,99 @@ impl Iterator for RepeatN {
         if self.remaining > 0 {
             self.remaining -= 1;
             Some(Output::Value(self.value.clone()))
+        } else {
+            None
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.remaining, Some(self.remaining))
+    }
+}
+
+/// An iterator that repeatedly yields the result of calling a function
+pub struct Generate {
+    function: Value,
+    vm: Vm,
+}
+
+impl Generate {
+    pub fn new(function: Value, vm: Vm) -> Self {
+        Self { function, vm }
+    }
+}
+
+impl KotoIterator for Generate {
+    fn make_copy(&self) -> ValueIterator {
+        let result = Self {
+            function: self.function.clone(),
+            vm: self.vm.spawn_shared_vm(),
+        };
+        ValueIterator::new(result)
+    }
+
+    fn might_have_side_effects(&self) -> bool {
+        true
+    }
+}
+
+impl Iterator for Generate {
+    type Item = Output;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let function = self.function.clone();
+        let result = match self.vm.run_function(function, CallArgs::None) {
+            Ok(result) => Output::Value(result),
+            Err(error) => Output::Error(error),
+        };
+        Some(result)
+    }
+}
+
+/// An iterator that yields the result of calling a function N times
+pub struct GenerateN {
+    remaining: usize,
+    function: Value,
+    vm: Vm,
+}
+
+impl GenerateN {
+    pub fn new(n: usize, function: Value, vm: Vm) -> Self {
+        Self {
+            remaining: n,
+            function,
+            vm,
+        }
+    }
+}
+
+impl KotoIterator for GenerateN {
+    fn make_copy(&self) -> ValueIterator {
+        let result = Self {
+            remaining: self.remaining,
+            function: self.function.clone(),
+            vm: self.vm.spawn_shared_vm(),
+        };
+        ValueIterator::new(result)
+    }
+
+    fn might_have_side_effects(&self) -> bool {
+        true
+    }
+}
+
+impl Iterator for GenerateN {
+    type Item = Output;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.remaining > 0 {
+            self.remaining -= 1;
+            let function = self.function.clone();
+            let result = match self.vm.run_function(function, CallArgs::None) {
+                Ok(result) => Output::Value(result),
+                Err(error) => Output::Error(error),
+            };
+            Some(result)
         } else {
             None
         }